+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2015-01-22
+## 2019-07-11
+- rewrite of the backend using http4s, tapir, monix and doobie
+
+## 2015-01-22
 - MongoDB replaced by Slick & H2 with Flyway for easy database schema management
 
-## [0.0.1] - 2014-12-16
-### Added
+## 2014-12-16
 - Adding mongo docker container start to backend-start.sh
 - Dependency version update
 - Adding changelog file